Adds mongoDB, posts to Madeleine, fetches & scrubs text from S3
This feature completes the cycle through the text process server
to the analysis server. The text server now fetches text data
from S3, scrubs the text, passes it to the analysis server, and
the analysis server posts to the madeleine with the sentiment
summary.

This also adds MongoDB and Mongoose as well as defines the
needed Schemas.